Basic working from single prog. No data transfered
